further extract dynamo/ecs interaction for deployment lambda
